requirements.txt
streamlit>=1.28.0
litellm>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# saving system
#faiss-cpu>=1.7.4
#sentence-transformers>=2.2.2 

#pypdf>=3.15.0
